Create README.md
I'm an aspiring Data Scientist, who loves to commit myself to the fullest of my dedication and commitment to my job. I am capable of turning the raw data into a meaningful conclusions and summary for arriving at the desired results with the help of various models of Python, AI, SQL, Power Bi and Advanced Excel.